+++ src/config/mod.rs
@@ -80,6 +80,7 @@ pub use generate::search::{SearchConfig, SearchField, SearchFormat};
 pub use hooks::HooksConfig;
 pub use html::anchors::{AnchorConfig, Place};
 pub use html::highlight::HighlightConfig;
+pub use html::region::RegionConfig;
 pub use html::{Footnotes, HtmlConfig};
 pub use lang::LanguageConfig;
 pub use links::external::ExternalConfig;